+++ src/client/components/board.jsx
