+++ src/client/components/column.jsx
